e-commerce-demo/requirements.txt
colorama
marqo
pandas
tqdm
python-dotenv
